fix problem with my_token feature in ip_profile_lib.py
It was unclear whether the user should put '?token=' in the token.conf file or not when specifying the user's token. Now the my_token feature is smart enough
to detect whether or not it is specified and add it in if needed.

Also, the my_token feature will identify unecessary whitespaces and remove them.